Add database configuration
Add configuration to specify the DB cluster and name where our tables
should be created.
In the installer/updater hook, only create tables if they should be
created locally (this check is not entirely accurate, but we can't do
much better than that).
Add a helper class that connects to the right DB.
And add boilerplate code for defining new services.

Change-Id: I93f97266bd2d73daaf4f117a02de38e91f6aff9c