Implement registration header & related features
This patch implements the registration header on event pages, with a
functional "More details" popup and functional buttons for
(un)registering.

An (incomplete) attempt to generate the popup in JS only was
made in PS 15, but there are too many limitations, so we just create it
in PHP and show it when needed.

Bug: T305697
Bug: T305699
Bug: T306647
Change-Id: Ief09dc778ade8e39150462a186b3fc998bf1d0a4