Prevent validation error with empty meeting URL
Null and '' have different meanings, the former meaning that no value
was passed, the latter meaning that a value was passed, but it's empty.
The latter is also validated, which was causing the form submission to
fail if no URL was specified, because it used the empty string instead
of null.

Change-Id: Ic2e95009017cc09248c04c209b742ae7a2230666