Stop passing Content-Type header in tests
As of core patch I9fb7f3c536af4636e33f776de32f9f90b88827c9 the body is
no longer ignored for DELETE requests, so if we pass a Content-Type we
also need to pass a (JSON) body. We don't need a body for those tests,
so just remove the header.

Change-Id: Ibe5f091e5116791c5b1979aa648a9752812d5bb1