Improve URL validation in EventFactory
PHP's FILTER_VALIDATE_URL does not enforce a specific scheme, and it
allows every character if it can't recognize the scheme. It also does
not accept relative URLs.

Fix this by adding some more checks and a test.

Note that as I wrote in the code, this is still subpar, and it's
actually quite weird that we have to do all the validation ourselves.

Change-Id: Id733f07e34b6b1c2b98d09c7bd7b02d0fd7e1286