Adjust CSS for the registration header and popup
Make it more similar to the wireframes by having even padding on the box
and adding some opacity to the icons. Also fix the aspect of the header
on very small screens where the content could overflow the container.

Make long words wrap if necessary, and make the popup columns have the
same width.

Also display the country in the popup.

Bug: T305697
Bug: T305699
Change-Id: If048c1f8b8235d546ec0627ac3ee66c0831dcf2f