Implement kebab menu and "edit event" action on Special:MyEvents
Other actions will be added in subsequent commits. Also reorganize the
relevant RL modules a bit; in particular, use the same module for
Special:MyEvents and the pager: we don't really expect the pager to be
used on its own, so it doesn't make sense to add two RL modules just for
it.

Bug: T310205
Change-Id: I9a8a7bc706e1424f2eeacb1166193c6c643a58db