Add the ability to search participants
Add a parameter to the "list participants" handler so clients can
specify a filter for usernames. The backend implementation is actually
subpar: it gets all participants from the DB and then filters the list
in PHP; unfortunately, there doesn't seem to be a clean alternative for
the moment.

Also fix the following bugs:
- Apply the width to the whole search widget, not just the input,
  otherwise the "x" button that appears when you type something will be
  outside the field;
- Set a fixed height on the user container so that it does not shrink
  and expand again when the list is rebuilt (e.g., when searching);
- In order for the above to work correctly when there are no
  participants, hide the user container if that's the case.
- Simplify code and output HTML a bit in EventDetailsParticipantsModule;

Bug: T308574
Change-Id: Id25581102ce853744f7751f91450addfe4f2b4ab