Fix conditional validation of meeting URL/address/country
In EventFactory, return an error if the URL was provided but the event
is not online, or if the address/country was provided but the event is
not in-person, instead of silently discarding the information, or saving
redundant data in the database.

Move validation to a separate method since it became quite long.

Avoid this error on the special page by filtering the input. Only API
clients should be able to see the new errors.

Bug: T312682
Change-Id: Ia0d8037f91ab6cf64def44d0e2c71aa5450da4f2