Enable some RL modules on mobile
These versions are not really optimized for mobile, but I tried them out
and they seem at least usable. Also, a non-optimized version is
certainly better than adding unstyled elements to the page (for
instance, event pages would be unreadable).

The eventpage module is not enabled on mobile because the basic
experience is already sufficient, so mobile users would get the no-JS
experience for now.

This is mostly intended as a stopgap until we have proper mobile styles.

Change-Id: I7b2f069e79b35e78e0a3ea4393d61b7cdd320b2c