Handle event page move and deletion
Also rewrite CampaignsPageFactory so that it doesn't validate the title
if it's stored in the database and also does not require it to exist.

Also add validation to ensure that the input does not contain a section
anchor (this was previously discarded silently).

Also, do not display the "enable registration" header/modal if the page
is a redirect.

Bug: T312775
Change-Id: Iae552fd2ff32c6a13f724a2c472f287abc2f3e0d