Make event address and country optional for in-person events
Both are now optional when creating or editing a registration, in both
the UI and the API endpoint. The display is changed to account for that,
for both the event page header and dialog, and the Special:EventDetails page.

Note that the result does not look that good when only the country or
the address is specified, but this is not really important right now
because we're going to have a single field once we start using the
geocoding service.

Bug: T313354
Change-Id: Ia534045e94b3a19a1f477e75397dfa3f67a8d771