Split user and authority
This patch introduces a new ICampaignsAuthority interface that replaces
ICampaignsUser in permission-related code. This new interface is similar
to core's Authority and represents the user performing the request. It's
also guaranteed to be local.

The ICampaignsUser interface is now only used for things unrelated to
permission checks, and it can represent global users.

The next step is to change the ICampaignsUser class and add the central
ID to it, so we can stop trying to convert central users to local ones.

Bug: T313133
Change-Id: I39f153ef99d482a1043461bf8ba142e7122b3e8e