Introduce CentralUser to replace ICampaignsUser
This patch is a fairly large refactoring of the concept of "user".
Before we would always convert global users to local users as soon as
possible, and convert local to global when doing things with the
database.

The new approach is that we'll (almost) never try to convert a global
user to local; instead, we'll use global users throughout. Local users
are converted to global as soon as possible.

Due to the change in the paradigm, some operations are now cheaper
(e.g., obtaining the central ID), and some are more expensive (e.g.,
getting the username).

I believe this is the right way to go, as it fixes a few issues we had
previously, as noted on the task.

This patch also adds some error handling for users without a global
account, and improves handling of deleted user thanks to the new logic.

Bug: T313133
Bug: T312772
Change-Id: I4d34ec16931e04eb8a4fbc1c05b452c7cae2919f