Improve behaviour of registration header and dialog
When the user (un)registers, reload the page so that the list of
participants is updated; this is easier than trying to update the dialog
in JS, as explained in the task.

This also means that we no longer need both "action items" and toggle
their visibility.

Also make sure that the current user is always included in the partial
list of participants, and ensure that they come first. Adjust the code
as needed to respect the limit of participants that can be shown here.

Bug: T312646
Change-Id: Ia06d15b188b73d63ff7bf1e73d4593ee9be479a4