Remove flatlist from location in enable registration form
Flatlist does not work well on small screens: the options will try to
fit in a single line but they'll eventually wrap, with an option
potentially spanning two lines.

Since the options are not guaranteed to fit on a single lines even on
desktop (the label may be longer in other languages), just remove
flatlist altogether.

Bug: T315496
Change-Id: If583aa2049b207297c03cbf7188502ed881c7995