Add timezone selector to Special:(Enable|Edit)EventRegistration
This uses the new time zone selector widget. For now, the timezone is
only used to convert the timestamp before saving, and discarded
immediately without being persisted. This preserves the status quo of
event times being in UTC, but will be changed soon. Another patch will
tweak the datetime inputs so that they work better with non-UTC
timezones.

Bug: T315691
Depends-On: I99a00dff7e3319ce45883191daee16bec1ed68ba
Change-Id: I98e3a8a36b08e851a897254171f42038e77eea0a