Make time inputs work correctly with non-UTC timezones
This patch tweaks the time input fields for start and end date so that
they work correctly when the timezone is not UTC. We first infuse the
fields manually so that we can specify the date format. Then, we make
sure that the default date, min, and max are updated when the timezone
changes. This is quite hacky and messes up with DateTimeInputWidget, but
it's much easier than adding proper timezone support to the widget.

Bug: T309631
Change-Id: Iea3adcd6fd847eef9641f5d3f60d78d1aef4145a