Update schema for tracking tools
Replace the old temporary columns with the new columns. Update the code
so that the names are consistent. Remove previous mentions to tracking
tools in commented code, they'll have to be rewritten anyway.

For now, only NULL can be inserted in these columns, as the logic for
actually handling tracking tools still has to be written (T312869).

!!!IMPORTANT: DO NOT +2 this patch. See instructions on the task.

Bug: T316409
Change-Id: I098d630f09416875797bccff589b40e123dc0a53