Create new columns to store local times and timezone
This patch creates the column to store the local timestamps and the
timezone of an event, and renames the existing columns so that it's
clear that they hold UTC values. For simplicity, this code writes
to the new schema but does not read the new non-UTC columns, so
that this patch can be more easily reviewed and applied on beta
(and reverted if need be). A subsequent patch will add the logic
for also reading the new columns, representing them in
EventRegistration objects, using the appropriate columns
everywhere, and exposing the values in the API endpoints.

!!!IMPORTANT: DO NOT +2 this patch. See instructions on the task.

Bug: T311126
Change-Id: I202f3470e8c79c5fd31b1e3ea8cb0cb7de097587