Introduce distinction between local and UTC time
The EventRegistration class now holds both, with the local time being
provided by the callers and the UTC time being computed by the getter
on-demand. All callers were adjusted to use the time that they actually
need. Note that we still always use the UTC time for the interface, but
that will change as part of T316688. The event timezone is always
assumed to be UTC, and it's not customizable yet.

Also, note that the UTC value is mostly internal (used e.g. for
sorting), so it is not exposed to clients and should not be used
whenever possible.

Tests were updated to verify that the new values are correct, and also
that the UTC conversion logic works as expected.

Bug: T311126
Change-Id: I1ed7e1e7997fcc28f12d666113c1129aa48569b0