Fix JS error when searching participants
onDeselectAll also unchecks the "select all" checkbox, which does not
exist when the current user is not an organizer. Thus, calling the
function when updating the search would result in an uncaught JS error.
Check whether the element exists before trying to access it, and do the
same in onSelectAll for consistency (even though right now that function
is never called if you're not an organizer).

Bug: T308574
Change-Id: I9cce223499add9eeb24d48746484220a141b6ffc