Refactor logic for showing a user link with fallbacks
Move the logic to UserLinker so it can be reused elsewhere. Move the
relevant styles to their own file for the same reason. Make the names of
the i18n message and CSS class more general.

Also handle the case of deleted users, with a similar message and style;
the color is slightly darker as a visual cue that this is different from
hidden users.

Also change the color of the placeholder for hidden users, as the
previous one did not pass WCAG AAA.

Finally, replace "removed" with "suppressed", which is possibly a
clearer indication of what happened and is in line with what Legal
suggested.

Bug: T316137
Bug: T318125
Bug: T318450
Change-Id: I31ef3b836aa5897c385d209bb29e10cb70cc166a