Rewrite FrontendModules to use proper DI
Create a factory for modules to avoid direct instantiation of complex
object, as well as avoid a long parameter list.

Move to the participants module some participants-specific code that dit
not belong to SpecialEventDetails.

This is still not perfect, and there's some space for improvement, but
should be a bit more maintainable than the status quo.

Change-Id: I9b1057e989773b2aaf89d629aa836d7f1a6b7351