Update schema for private registration
Add the new field to the table, and update write queries to account for
it. In theory there'd be no need to change read queries now, but I've
done it anyway to make sure that we will remember to update them again
once we build actual support for private registration.

Note that the column is not indexed because it has low cardinality; we
expect private registrants to be less than public registrants, but we
never query private-only, so an index shouldn't really make any
difference.

!!!IMPORTANT: DO NOT +2 this patch. See instructions on the task.

Bug: T318120
Change-Id: I5b96981397eea4d66a9d729dfb309b4d1ffcd5cc