Add tests to document UTC conversion of ambiguous times
This test documents the behaviour of the application when the local time
is ambiguous (skipped or repeated).

Skipped times are retained as entered by the organizer; I'm not sure if
this is fine, but should be OK for now.

For everything else, we just rely on PHP to do the right thing.

Change-Id: I72a76ed12afb9f78cf2c17ed190f85c5e3e55c3b