Fix kebab menu options on firefox
Unlike chrome, firefox will follow the href when an option is clicked,
so it will redirect to the target page even if the API call succeeds. In
fact, I'm not even sure why it worked in chrome in the first place.

So add a listener for the left mouse button and explicitly prevent the
default event handling.

Bug: T321620
Change-Id: Id78edb2857861879cc2f93d39cce52297a640ccf