Introduce EventTimeFormatter
This service is responsible for formatting the event dates and time zone.

For now it simply does what the previous code was doing, but this will
be changed in a subsequent patch. The time correction is chosen and
passed in explicitly so that it will be easier to see the difference.

::formatTimezone() is also currently unused, and it ignores the $event
parameter, but it was added for completeness and because we will need it
soon.

The formatted time is encapsulated in a value object so that we can
access every representation without having 3 distinct methods in the
formatter.

Bug: T322054
Change-Id: I05917ee6dacefc6eb2ff969193048aeaf7278d1b