Refactor ParticipantsModule to handle the current user separately
Separate the current user from the list of other participants, pass both
parameters around, and use a separate method to render the current
user's row. Also simplify the code a little bit.

For now, this is needed because the current user's row need some special
handling to fix T319453#8372310. In the future we may want to add some
controls next to the current user's name, so we'd need a separate method
anyway.

Change-Id: Ie43c17ed5f04a0be367cb58cbc764eced513731e