Special-case the current user's row on EventDetails when searching
This fixes the bugs reported in T319453#8372310. The current user is now
special-cased when searching; its visibility is determined in JS based
on the username filter, and the row is never deleted, only hidden or
shown.

Also refactor this code a little bit, since I think
deleteParticipantsList() was very special and it should generally not be
used except for when searching, so make the method more specific and
move there some more statements as well.

Also avoid the usernameFilter class property, as it can be passed in
directly.

Bug: T319453
Change-Id: I3540c2e87c9924ee895eb563f7ed7798c695431f