Introduce TrackingToolEventWatcher to validate tracking tool changes
This patch implements an event watcher that reacts to changes to an
event (creation, update, deletion, change to the participant list) and
informs any attached tracking tools of said change.

For now, the action can only be "validated", in what could be thought of
as the first phase of a 2-phase commit. Another patch will implement the
actual execution of said actions in the tracking tools; this work is
tracked in T320434.

Note that the concrete implementation of this logic for the WikiEdu
dashboard is still just a stub meant to demonstrate what things will
look like. Writing the actual logic is tracked in T317707.

Bug: T331884
Change-Id: Id389bcbcda716cd2425598341a34686db8ce553d