Add new ce_tracking_tools table
This patch introduces a new table for storing tracking tools, as
described on the task. The type of the tool_event_id field is also
changed from blob to a 512-bytes varbinary, since IDs are expected to be
short.

Note that the migration logic reads the global state, instead of using
Config. This is for simplicity, since the flag will be removed soon.

Some parts of the logic added in this patch are still subject to change,
like what part of the code will actually be responsible for updating
tracking tools. Since it's not possible to add any tracking tools yet,
the code can be changed very easily. The current implementation provides
a proof of concept, as the important thing for now is creating the table.

Bug: T334143
Change-Id: I47030c94895e7bcdf7e9b820e49cd67e631d8dc0