Decompose JOINs on address tables in EventStore
Use separate queries instead of a single queries with multiple joins:
this is a many-to-many relationship, and the result set could grow a
lot. That becomes an even bigger problem if we add new tables to store
other information about events, e.g., tracking tools (T334143). DBAs
recommended this approach because running additional queries does not
impact performance, and this is better than complex queries.

Addresses are retrieved in batch to avoid the N+1 problem, and grouped
by event ID later in the code.

Also fix an issue where batch queries would apply LIMIT to the whole
result set, and not just the events.

Bug: T334483
Change-Id: Ie5048298314a762ce3bcc4939b09576271ad4986