Fix failing selenium test
Selenium tests are failing on master because a test used 2023-04-15
as the start date, but that's now in the past. Instead, make sure to
use dates in the future, by using the year 2999. Also abort
the test if the API request fails, instead of trying to use the API
error as event ID. As a consequence of this, other errors that were
previously hidden have now become apparent:

- Rewrite the methods for changing the start/end date: we need to
blur the input after updating the year, otherwise the change won't be
recorded.
- Move tests for participant registration to their own file, and enable
  the registration only once in that file, instead of once per test
case, which would fail because the page already has a registration
associated.

Also:
- Use different page names prefixes for different tests, so it's easier
  to figure out what test used what page.
- Use Util.getTestString instead of redefining the method.
- Rename createEvent to createEventPage to avoid any confusion with
  event registrations. I think that method would ideally belong to a
different page objects, but it's fine for now.

Change-Id: I45914190f951068c5cc7bc8624d6b424f25844db