Add participant questions on RegisterForEvent special page
This patch also implements the question registry, and a method for
obtaining HTMLForm descriptors for the questions. This is behind a
temporary feature flag.

It also introduces a separate flag to enable Wikimedia-specific
questions like "confidence contributing" and "affiliate". In the future,
it might be better to have a hook for adding custom questions, and
handle that hook in the WikimediaMessages extension. However, since this
code is still very much subject to change and the database part of it
still has to be implemented, it's better to keep all questions in the
same place for now.

Bug: T335808
Change-Id: I9d8bf981169871a96ba27a1f5a11ac3a6c061230