Handle PageDeleteCompleteHook instead of PageDeleteHook
This guarantees that the event is only deleted if the event page
deletion succeeds, whereas previously this wasn't the case.

Handling this hook means that we can no longer abort the deletion. As
such, skip all the validation of tracking tools to ensure that the event
deletion succeeds. Also, the deleting user likely has no control over
the tracking tool that would be blocking deletion, and if the tool has
some sort of error, this would result in a page impossible to delete.

This patch is a partial revert of
I72ddd29a54fd86236d29b7b926349a27e8fb4760.

Bug: T337946
Change-Id: Id8958329ed92dad72807d4f6f086cef92d718e83