Sync participants when linking a WikiEduDashboard to an existing event
If the event is created but not linked to any tool, then someone
registers, and then the organizer links the event to the dashboard, we
want to make sure that existing participants are immediately
synchronized. Since the dashboard doesn't do that for us when syncing a
new event, we need to do it manually.

Bug: T317707
Change-Id: Iaa895594723959907628c147e5a71ad627a07c78