Make it possible to change visibility in Special:RegisterForEvent
Users who are already registered for an event can now use the special
page to edit their registration visibility.

The old error message is now unused. The form legend changes depending
on whether the user is already registered. Same for the success message
and submit button.

The "confirm" field was removed because it's no longer necessary now
that the form actually has fields.

Return more information from ParticipantsStore about the result of the
change, so that it's possible e.g. to not send a new confirmation email
when the user is only changing their visibility.

Always show the form after submission (with a success message on top),
so that the participant can immediately see the updated data.

Bug: T319454
Change-Id: If60105d5f5b49c552fafc175eda118c1b8c36f11