Rename RegistrationConfirmationDialog and related code
This is no longer a simple confirmation dialog as it used to be. It
already has a field for setting the visibility, and soon it'll have
fields for participant questions too. Also, it can be used when editing
your registration, which is not really a confirmation of anything.

Update the dialog title and button label to match the latest design in
T319454.

Some i18n keys are intentionally left untouched, as changing them is
more complex.

Change-Id: I3a8eb489f9e6760519a09e56d1ca668050eab14f