Add support for temporary users
Every time we check whether the user is registered, what we really want
to know is whether they're named: the extension intends to treat all
temporary users the same as logged-out users.

So replace the various isRegistered checks with new isNamed checks,
which verify that the user is registered and non-temp.

Also fix a weirdness in the canDeleteRegistrations check where we were
potentially allowing logged-out users to delete a registration if they
had the user right.

Bug: T326872
Change-Id: I54bebf53a562f6dfbd5c7cb1f9423b48b792db4a