Make OrganizerSelectionFieldEnhancer use its own API handle
Sharing the same handle as the UsersMultiselectField leads to an
unwanted interaction where as a tag is chosen, we start an API request,
but while that's on flight the value of the textinput is reset to the
empty string, which causes UsersMultiselectWidget.updateMenuItems to be
called, which in turn aborts all pending API requests. Since we share
the same handle, it'll also abort our request, meaning the new value is
not validated.

Bug: T338635
Change-Id: I163ac054bc7eb1cec72d74128db5e23375993a20