Add fields to the registration form for tracking tools
Organizers are now able to connect the P&E Dashboard from
Special:EnableEventRegistration and Special:EditEventRegistration. Note
that the implementation here is not tool-agnostic; instead, it bery much
assumes that the P&E Dashboard is the only available tool. This should
be fine for now though, and can be improved later.

Note that the field, unlike the API, accepts a course URL. This means
that we need additional validation and URL parsing to make sure that the
provided URL is valid, and tool classes are responsible for this.

Bug: T317708
Change-Id: I02ffa5d3954c54c2831807573fdd80d94dd0c40a