Fix crash with event ID when syncing a new event with the Dashboard
When syncing a new event, the $event object doesn't have an ID because
the object itself is not updated after the DB insertion. This would
currently result in an uncaught LogicException if you try to add the
Dashboard when creating a new event.

So add a separate integer parameter to guarantee that an integer ID is
always passed when syncing a new event, rather than throwing an
exception later on.

Replace addToEvent with two methods, one for existing events and one for
new events, with different preconditions and type checks.

The event ID continues to be optional for validation (i.e., dry_run), as
it won't be available at that point for new events.

Bug: T317707
Change-Id: I3d64f0ce82191e5f6e39c4be5e5a204fea6d5282