Try to make the myEvents selenium spec more robust
This spec is failing relatively often, see for instance all the failed
gate-and-submit attempts in I3bcc1654a181444619735d6c07b83b2aa9569624.

I'm not really sure why that's the case, but at least this patch makes
the spec more robust by filtering for open events first, so as to make
sure that the first event in the list is open, and that it can therefore
be closed.

Change-Id: I70d5a1e1a563095dc4310143f5323c325c62a8d7