Create a backend for adding questions to an event
This patch implements the logic necessary for the ce_event_questions
table, i.e., storing what questions are enabled for an event.

The list of questions is stored in EventRegistration objects, so that
it's easier to retrieve elsewhere. EventStore reads and writes this
list, delegating to the new EventQuestionsStore.

When an event is created, we now enable all questions for it, as decided
for the MVP (and as long as the feature flag is enabled). When an event
is updated, we retain the previous questions.

Note that this patch is only for the "organizer side", i.e., it does not
affect what participants see. This will be done in a subsequent patch.

Bug: T340125
Change-Id: Ieec43f95530abc91f386af24baf2dda17af225da