Create backend for storing participant answers
This patch introduces ParticipantAnswersStore, which is the data access
object for participants answer (ce_question_answers table).

It also introduces Answer, a new value object for representing said
answers.

The new store object is functional and wired up to ParticipantsStore,
but only for the "write" part. A method for reading answers is provided,
but that is not wired up yet, and it will be in a subsequent patch.

ParticipantsStore itself now supports setting answers, but its callers
always pass an empty array, thus preserving the status quo. This will
also be changed in a subsequent patch.

Bug: T339259
Bug: T339257
Change-Id: Ief97dd88a1b6b4287d12dafceb58345feb13d1b7