Store participant answers when registering
Add an $answer parameter to the high-level methods, and pass it down to
ParticipantsStore. This is behind the feature flag and for now it's just
a partial implementation:
- The API doesn't allow answering any questions -- all API requests will
  leave the previous answers unchanged.
- The special page doesn't support editing your answers yet, so in that
  case we will also leave the previous answers unchanged.
- When registering via the special page, we will save the answers. Note
  that the code can't use HTMLForm at the moment (instead, it needs to
  interact with the WebRequest directly) because the questions are not
  shown when editing your registration info, and therefore they're not
  provided to us by HTMLForm.

Add a requirement that all multiple choice questions must have a 0
option used as a placeholder, and ignore placeholder values when parsing
the form data.

Because the fallback to the current answers (in special page for edits
and API endpoint) is temporary, it accesses the needed services via the
global service locator directly.

Bug: T339259
Change-Id: I518de84880701dfcb5fe4eb606a1b81a0ad76cd3