Add columns to ce_participants for participant questions
Add the timestamp of the first answer and that of the aggregation, as
explained in the task. Note that because both columns are DEFAULT NULL,
this schema change is backwards-compatible.

The new columns are nonetheless added explicitly to existing INSERT
statements in tests for clarity. Those in non-test code are commented
out for the moment, and until the new columns are created in production.

Bug: T339982
Change-Id: Ibd7b4cb1bef00350a79bf911e93c58ca146ace30