Add participant questions support to API: get possible answers
Add an API endpoint that returns a list of all available answers,
potentially filtered by ID. This uses a format that's more useful for
API users; for instance, it provides localisation, although it needs to
use the content language because of T269492.

Also make it possible to replace the questions in tests, to avoid
depending too much on the implementation, and to have shorter test
cases.

Bug: T340736
Change-Id: If5b945cc22683ade1042c3451fb762acf65133d9