Fix participants pagination when searching and scrolling
When we rebuild the list a scroll event is triggered, which causes the
first batch of participants to be loaded twice (once from the "scroll",
and once for the list rebuild).

Add a method to pause the observer when rebuilding the list, and unpause
it afterwards once the DOM's been redrawn.

Bug: T340897
Change-Id: I05cdf7d532b716ec355d0c54dffcb0ea89e931b1