Update format of participant answers returned by the API
Database IDs are an internal implementation detail, expose the names
instead. The option+text division is also internal, so use value+other.

Do the formatting inside the registry, which will also be responsible
for reading an array of answers in the same format and transform it into
an array of Answer objects for us to store.

Bug: T340736
Change-Id: I159c75eb578ed840cbb7d1462fdfda1aba1f8e65