Special:EventDetails: Fix participants search on mobile
Allow the flex container to wrap, so that it can put each item on its
own line instead of squeezing them in a single one. Adjust spacing as
needed. This can certainly be improved, but the latest prototype is
entirely different anyway, so this should suffice as a temporary
solution.

Change-Id: I8166622649bcc525e22634ecab1c1643ded2787b