Remove default value from campaign_events.event_answers_status
Remove the default value and add the column explicitly to the write
queries, so that the code will work regardless of whether the default is
there. Note that for now the code is just using a plain 0, but in the
future this should use a constant, and of course it should insert the
appropriate value. This patch is intentionally kept small to reduce the
chances of anything breaking.

Bug: T341142
Change-Id: I2780a8df08978f13b3597f03d55421c7ef2845a1