Avoid flashing "edit registration" buttons on event page if JS enabled
Hide the no-js buttons if JavaScript is enabled, so that they won't
flash for a split second (or more, if the network/PC is slow) before
being replaced by a ButtonMenuSelect. This way, JS users won't initially
see anything, and then the JS version will appear in the blank space at
the end of the header.

Bug: T319454
Change-Id: I936984eea31885befd7076c7f5a35e045a1f3e32