Make EventQuestions.js use the API format, not HTMLForm
This is not an HTMLForm, so if we try to use the HTMLForm format we're
forced to reverse-engineer it, going into the internals of HTMLForm.

Instead, use the format for API, which is meant for clients. In fact, we
could even do all this in the client side with some API calls, but for
now this is fine (and probably more efficient, especially considering
the number of messages potentially used here).

Doing this implies that EventQuestions needs to shift from a "hide-if"
paradigm to "showIfVal", which is implemented in the new version. Other
things like CSS classes are now done directly in JavaScript.

Also fix name of extractUserQuestionsAPI: the method really extracts
answers, not questions.

Change-Id: If1a8a6cf75ba03a98bef41e4fe3a204ec6b771b8