Restore message documentation deleted due to conflict
In I3dddea3f, I accidentally introduced a duplicate message. This isn't
allowed, and L10n-bot got confused: first, it updated the non-API
message documentation with the API documentation in I4f95e6a8. Then I
deduplicated the messages in I9610b11c, but the bot later removed the
documentation from the non-API message in I5084fced. This resulted in
failing CI on master.

Simply re-add the documentation with the content it had in the original
patch.

Change-Id: I54a580166998c0a914872c7aa79bf0f64c864db9