Introduce a maintenance script for aggregating participant answers
The script scans the ce_question_answers table to find answers to
aggregate (based on event end date and participant registration
timestamp); it moves the aggregated answers to the
ce_question_aggregation table and deletes the original rows. Finally, it
updates cep_aggregation_timestamp for each user.

All this is done in a single transaction to guarantee atomicity and
isolation. The performance may not be ideal, but I think this is needed
due to the nature of the data. Also, these tables are pretty small, and
we can think about performance again if they get larger.

Bug: T336858
Change-Id: I731b477253e797f629225ae26232c0c42bfba548