Add stats tab with empty states to EventDetails
Add the "response statistics" tab, which is only shown after the event
has ended. Implement the two empty states for when there are no answers,
and for when answers still have to be aggregated.

The tab only shows some placeholder text in place of the actual data
when answers have already been aggregated.

Bug: T341341
Change-Id: I76884f7c8f80c9e7b90ed8d9995df7409a012d6a