UserLinker: avoid expensive existsLocally lookup
The result of the method calls is never used, because the two branches
do the same thing. But the method is quite expensive, taking up an
inclusive 18% of the time for a simple GET request to the
ListParticipantsHandler endpoint in production. Locally, I get a 13%
speedup with this patch.

Also clean up some TODOs and documentation.

Bug: T345250
Change-Id: I5339b5895679f2d92e1a87570c8a67a6c33ecca0