Make ListParticipantsHandler lookup all usernames together
Iterate over the list of participants twice; the first time, we only
build a map of user IDs. That map is then converted to a list of
usernames (or metadata for suppressed/deleted users) with a single
query. This map is then used in the second iteration to obtain the
username of each participant.

The response is generated roughly 8% faster locally.

Bug: T345250
Change-Id: Ib1bb7533fa512965a72e501d6cf62a2c769f3645