Preload user page link data in ListParticipantsHandler
Add a new method to UserLinker to preload link data for a batch of
users, essentially a wrapper around core's LinkBatch. This allows the
Linker code to run a single query to see which ones of the given user
pages exist, instead of running a separate query for each page.

Locally, the response is generated 5% faster after this patch.

Bug: T345250
Change-Id: I739e17c09c301b818bac047298daea3ca2108a96