Fix flaky selenium spec eventPage.js
Sometimes the test will attempt to click the "Register" button in the
dialog before the dialog is actually ready. I could reproduce this quite
reliably locally, and it seems to be failing a lot in CI lately.

Add a waitUntil() to make sure that we don't try to click anything until
the dialog and its buttons are ready.

Change-Id: I6829e0cf523d782361497e6bfcfbde0b86c2743e