Add username caching to CampaignsCentralUserLookup
Add a cache of ID => usernames, so that multiple calls to getUserName /
getNames won't result in multiple DB queries. For instance, this is very
important in the ListParticipantsHandler class, which runs a single
lookup for all the participant, but then calls
UserLinker::getUserPagePath on each one of them, which results in
duplicated getUserName calls.

Also add some tests to verify that these methods and the caching work
correctly.

Bug: T345250
Change-Id: I585b26957c0ec80e83e39deb40e6b2c52f2517f0