Send registration notifications in a deferred update
Sending the notification is slow, because it need to generate the email
and then send it. By making it non-blocking, the main request can
complete much sooner.

According to an excimer profile [1] I ran on beta (so possibly not very
reliable), notifyRegistration() alone takes up 41% of the time for a
full Special:RegisterForEvent submission (and probably something similar
for an API request).

A similar run on testwiki [2] says it spent 14% of the time there
instead.

The latter is probably more accurate, but whatever the actual value
might be, this change should definitely speed things up noticeably.

[1] - https://performance.wikimedia.beta.wmflabs.org/excimer/profile/937a7a32c19444cb
[2] - https://performance.wikimedia.org/excimer/profile/9d8c0ebf4d1846dd

Change-Id: I35c8abc12ef66c9fb87a844aa74592608666388b