Use native types for EmailUsersJob parameters
All job parameters should be JSON-serializable, and MailAddress is not
(T346406). Instead of using MailAddress directly, pass an array with of
strings that can be used to reconstruct the MailAddress object.

This fixes a bug where the job would not run in beta (and likely
production), possibly because the object parameters are not handled
correctly.

Also improve phan annotations and add a test to make sure this won't
happen again.

And finally, make $replyTo nullable and only pass it to UserMailer when
not null.

Bug: T346403
Change-Id: Ib6e14bb7e4fe75ed5e1e169a4925ad32417e080a