Hide retention info in Special:RegisterForEvent if the event has no questions
If the event has no questions, there's nothing for the user to answer,
and it would therefore be misleading to say things like "your
responses" or "edit ... anytime before" if the user can't do anything
with it.

Note that the feature flag check is left in the caller method to
minimise the diff (and consequent git blame effect) of its upcoming
removal.

Bug: T348276
Change-Id: If7de7bcf2f6856d5f48183b1c9c112393f480536