Adjust min dates in the registration form according to the time zone
The HTMLForm fields for dates can only operate in UTC, and do not know
about our time zone selector. This can lead to bugs when validating
event dates against the minimum allowed date, because the two might be
using different time zones.

So, add a workaround where we dynamically set the minimum allowed date
accounting for the selected time zone, if any. Fall back to UTC if no
time zone was submitted, or if it's invalid.

Note that this was already handled in the client side (see
TimeFieldsEnhancer.js), it's just the server-side validation that was
off.

The proper solution to this would be to have the time fields accept a
time zone parameter and let them do the calculation, but that's much
more complex.

Bug: T348579
Change-Id: I4f31440f2477c99d655d4226333d3a850a1079cb